3.0.7

+++ src/main/java/com/robertx22/mine_and_slash/database/quests/data/QuestTaskData.java
@@ -40,9 +40,9 @@ public class QuestTaskData {
             }
         }
     }
-
-    public int getCurrentCompletition() {
-        return (int) this.currentCompletition;
+    
+    public float getCurrentCompletition() {
+        return this.currentCompletition;
     }
 
     public void increaseCompletition(float add) {
